# web/app/docs/swagger.yaml
basePath: /
definitions:
  res.Error:
    properties:
      message:
        example: 'got a error in server '
        type: string
      status:
        example: error
        type: string
    type: object
  res.ValidationError:
    properties:
      field_name:
        example: subject_id
        type: string
      message:
        example: should be a integer
        type: string
    type: object
  viewip.edge:
    properties:
      id:
        type: string
      map:
        type: integer
      relation:
        type: string
      source:
        type: integer
      target:
        type: integer
    type: object
  viewip.node:
    properties:
      id:
        type: integer
      image:
        description: = 'lain.bgm.tv/img/no_icon_subject.png'
        type: string
      info:
        type: object
      map:
        type: integer
      name:
        type: string
      name_cn:
        type: string
      subject_id:
        type: integer
      subject_type:
        type: string
    type: object
  viewip.subjectMap:
    properties:
      edges:
        items:
          $ref: '#/definitions/viewip.edge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/viewip.node'
        type: array
    type: object
  viewip.subjectMapRes:
    properties:
      data:
        $ref: '#/definitions/viewip.subjectMap'
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
host: api.bgm38.com
info:
  contact:
    email: i@trim21.me
    name: Trim21
    url: https://github.com/Trim21
  description: A Set of http api for bangumi.
  license:
    name: MIT
    url: https://github.com/Trim21/bgm38/blob/dev/LICENSE
  termsOfService: "No"
  title: bgm38 api server
  version: 0.0.1
paths:
  /bgm.tv/v1/calendar/{user_id}:
    get:
      description: |-
        根据在看的番剧生成ics格式日历
        如果浏览器访问时会返回纯文本数据
        在使用日历app导入时会返回日历数据
      operationId: watchingCalendarV1
      parameters:
      - description: user_id, 可以在个人主页的网址中找到
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - text/calendar
      - application/json
      responses:
        "200":
          description: text/calendar
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/res.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.ValidationError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/res.Error'
      summary: generate a calendar from watching collection
  /bgm.tv/v2/view_ip/subject/{subject_id}:
    get:
      description: Get related Subjects and their relations
      operationId: viewIP
      parameters:
      - description: Account ID
        in: path
        name: subject_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/viewip.subjectMapRes'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/res.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Error'
      summary: get subject relations map
schemes:
- https
swagger: "2.0"